View Models refactored
