+++ VultisigApp/VultisigApp/Extensions/StringExtension.swift
@@ -40,20 +40,6 @@ extension String {
     var nilIfEmpty: String? {
         return isEmpty ? nil : self
     }
-    
-#if os(iOS)
-    func widthOfString(usingFont font: UIFont) -> CGFloat {
-        let fontAttributes = [NSAttributedString.Key.font: font]
-        let size = self.size(withAttributes: fontAttributes)
-        return size.width
-    }
-#elseif os(macOS)
-    func widthOfString(usingFont font: NSFont) -> CGFloat {
-        let fontAttributes = [NSAttributedString.Key.font: font]
-        let size = self.size(withAttributes: fontAttributes)
-        return size.width
-    }
-#endif
 }
 
 // MARK: - String constants
