+++ VultisigApp/VultisigApp/Services/Actions/Coin+ChainAction.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 extension Chain {
-    
     var defaultActions: [CoinAction] {
         let actions: [CoinAction]
         switch self {
@@ -60,17 +59,5 @@ extension Chain {
             actions = [.send, .memo]
         }
         return actions.filtered
-    }
-}
-
-extension Array where Element == CoinAction {
-    var filtered: [CoinAction] {
-#if os(macOS)
-        return self
-#elseif DEBUG
-        return self
-#else
-        return filter { $0 != .swap && $0 != .memo }
-#endif
     }
 }
