+++ VultisigApp/VultisigApp/Utils/Styles/ColorStyle.swift
@@ -8,17 +8,6 @@
 import SwiftUI
 
 extension Color {
-    // Font
-#if os(iOS)
-    static let systemFill = Color(UIColor.systemFill)
-    static let secondarySystemGroupedBackground = Color(UIColor.secondarySystemGroupedBackground)
-    static let systemBackground = Color(UIColor.systemBackground)
-#elseif os(macOS)
-    static let systemFill = Color(NSColor.systemFill)
-    static let secondarySystemGroupedBackground = Color(NSColor.controlBackgroundColor)
-    static let systemBackground = Color(NSColor.windowBackgroundColor)
-#endif
-    
     static let gray500 = Color(red: 0.96, green: 0.96, blue: 0.96)
     static let gray400 = Color(red: 0.92, green: 0.92, blue: 0.93)
