+++ VultisigApp/VultisigApp/iOS/Extension/ColorStyle+iOS.swift
@@ -1,0 +1,16 @@
+//
+//  ColorStyle+iOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-27.
+//
+
+#if os(iOS)
+import SwiftUI
+
+extension Color {
+    static let systemFill = Color(UIColor.systemFill)
+    static let secondarySystemGroupedBackground = Color(UIColor.secondarySystemGroupedBackground)
+    static let systemBackground = Color(UIColor.systemBackground)
+}
+#endif
