+++ VultisigApp/VultisigApp/iOS/Extension/StringExtension+iOS.swift
@@ -1,0 +1,18 @@
+//
+//  StringExtension+iOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-27.
+//
+
+#if os(iOS)
+import SwiftUI
+
+extension String {
+    func widthOfString(usingFont font: UIFont) -> CGFloat {
+        let fontAttributes = [NSAttributedString.Key.font: font]
+        let size = self.size(withAttributes: fontAttributes)
+        return size.width
+    }
+}
+#endif
