+++ VultisigApp/VultisigApp/iOS/Utils/Coin+ChainAction+iOS.swift
@@ -1,0 +1,20 @@
+//
+//  Coin+ChainAction+iOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-27.
+//
+
+#if os(iOS)
+import SwiftUI
+
+extension Array where Element == CoinAction {
+    var filtered: [CoinAction] {
+#if DEBUG
+        return self
+#else
+        return filter { $0 != .swap && $0 != .memo }
+#endif
+    }
+}
+#endif
