+++ VultisigApp/VultisigApp/iOS/View Model/MacCameraServiceViewModel+iOS.swift
@@ -1,0 +1,13 @@
+//
+//  MacCameraServiceViewModel+iOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-27.
+//
+
+#if os(iOS)
+import AVFoundation
+
+@MainActor
+class MacCameraServiceViewModel: NSObject, ObservableObject { }
+#endif
