+++ VultisigApp/VultisigApp/macOS/Extension/StringExtension+macOS.swift
@@ -1,0 +1,18 @@
+//
+//  StringExtension+macOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-27.
+//
+
+#if os(macOS)
+import SwiftUI
+
+extension String {
+    func widthOfString(usingFont font: NSFont) -> CGFloat {
+        let fontAttributes = [NSAttributedString.Key.font: font]
+        let size = self.size(withAttributes: fontAttributes)
+        return size.width
+    }
+}
+#endif
