+++ VultisigApp/VultisigApp/macOS/View Model/MacCameraServiceViewModel+macOS.swift
@@ -1,18 +1,16 @@
 //
-//  MacCameraServiceViewModel.swift
+//  MacCameraServiceViewModel+macOS.swift
 //  VultisigApp
 //
 //  Created by Amol Kumar on 2024-08-15.
 //
 
-import AVFoundation
 #if os(macOS)
+import AVFoundation
 import AppKit
-#endif
 
 @MainActor
 class MacCameraServiceViewModel: NSObject, ObservableObject {
-#if os(macOS)
     @Published var showCamera = false
     @Published var detectedQRCode: String?
     @Published var isCameraUnavailable = false
@@ -95,10 +93,8 @@ class MacCameraServiceViewModel: NSObject, ObservableObject {
     func getSession() -> AVCaptureSession? {
         return session
     }
-#endif
 }
 
-#if os(macOS)
 extension MacCameraServiceViewModel: AVCaptureVideoDataOutputSampleBufferDelegate {
     func captureOutput(_ output: AVCaptureOutput, didOutput sampleBuffer: CMSampleBuffer, from connection: AVCaptureConnection) {
         guard let pixelBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else { return }
